requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
pillow==11.0.0
PyJWT==2.10.0
pytz==2024.2
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
